useNumpadDecimal hook for numeric fields
This commits brings back a behavior that was
present on old fields that inherited from
FieldNumeric on which a keypress on the numpad
decimal separator key (usually '.') replaced
the key to the actual localized separator
character
