+++ addons/web/static/src/fields/float_field.xml
@@ -3,7 +3,7 @@
 
     <t t-name="web.FloatField" owl="1">
         <span t-if="props.readonly" t-esc="getFormattedValue()" />
-        <input t-else="" t-att-type="props.inputType" inputmode="numeric" class="o_input" t-att-value="defaultInputValue" t-on-change="onChange" t-att-step="props.step" />
+        <input t-else="" t-ref="numpadDecimal" t-att-type="props.inputType" inputmode="numeric" class="o_input" t-att-value="defaultInputValue" t-on-change="onChange" t-att-step="props.step" />
     </t>
 
 </templates>
