+++ addons/web/static/src/fields/integer_field.xml
@@ -3,7 +3,7 @@
 
     <t t-name="web.IntegerField" owl="1">
         <span t-if="props.readonly" t-esc="props.format(props.value)" />
-        <input t-att-type="props.inputType" inputmode="numeric" t-else="" class="o_input" t-att-value="formattedInputValue" t-on-change="onChange" t-att-step="props.step" />
+        <input t-ref="numpadDecimal" t-att-type="props.inputType" inputmode="numeric" t-else="" class="o_input" t-att-value="formattedInputValue" t-on-change="onChange" t-att-step="props.step" />
     </t>
 
 </templates>
