+++ addons/web/static/src/fields/percentage_field.xml
@@ -8,6 +8,7 @@
         <t t-else="">
             <div class="d-flex">
                 <input
+                    t-ref="numpadDecimal"
                     t-attf-class="o_input"
                     type="text"
                     inputmode="numeric"
